Added lots of new RIF built-ins. Updated hierarchy of numeric terms in order to be compatible with XSD. All numeric terms now use BigDecimal in order to allow arbitrary-precision values for specific data types.

+++ iris/trunk/src/org/deri/iris/builtins/string/StringSubstringUntilEndBuiltin.java
@@ -62,7 +62,7 @@ public class StringSubstringUntilEndBuiltin extends FunctionalBuiltin {
 			IIntegerTerm beginIndex = (IIntegerTerm) terms[1];
 
 			String substring = string.getValue().substring(
-					beginIndex.getValue());
+					beginIndex.getValue().intValue());
 
 			return Factory.TERM.createString(substring);
 		}
